+++ CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres roughly to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 ### Added
+- Column attribute `funcname` holds the name of the function call (if any). e.g. `sum`
+- Column method `aggregate_function?` returns true if the column is a sum, count, min, max or avg aggregate
+- Report method `summarised_query?` returns true if the report has a GROUP BY clause
+- Report method `non_aggregate_columns` returns an array of columns that are not aggregates
+- Report method `aggregate_columns` returns an array of columns that are aggregates
+- Report method `change_column_selection` takes arrays of non-aggregate and aggregate columns and removes any not included from the report's columns list and the group by clause
+- Report sets the `summarised` attribute when represented in a hash
 ### Changed
 ### Fixed